Move matrix utils to their own directory
