+++ server/lib/matrix-utils/ensure-room-joined.js
@@ -3,9 +3,9 @@
 const assert = require('assert');
 const urlJoin = require('url-join');
 
-const { fetchEndpointAsJson } = require('./lib/fetch-endpoint');
+const { fetchEndpointAsJson } = require('../fetch-endpoint');
 
-const config = require('./lib/config');
+const config = require('../config');
 const matrixServerUrl = config.get('matrixServerUrl');
 assert(matrixServerUrl);
